extract test suite
