+++ ydb/core/persqueue/pqtablet/partition/mlp/ut/ya.make
@@ -7,6 +7,7 @@ SIZE(MEDIUM)
 
 SRCS(
     mlp_commit_ut.cpp
+    mlp_consumer_split_ut.cpp
     mlp_consumer_ut.cpp
     mlp_counters_ut.cpp
     mlp_dlq_mover_ut.cpp
